pyproject.toml
[project]
name = "model-iq-tech-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    "fastapi>=0.121.3",
    "httpx>=0.28.1",
    "pydantic>=2.12.4",
    "pydantic-ai>=1.20.0",
    "pydantic-settings>=2.12.0",
    "pymongo>=4.15.4",
    "pytest>=9.0.1",
    "python-dotenv>=1.2.1",
    "structlog>=25.5.0",
    "tiktoken>=0.12.0",
    "uvicorn>=0.38.0",
    "pytest-asyncio>=0.23",
    "anyio>=4.2",
    "google-generativeai>=0.8.5",
    "pinecone-client>=6.0.0",
    "pinecone>=8.0.0",
]

[dependency-groups]
dev = [
    "black>=25.11.0",
    "mypy>=1.18.2",
    "pytest>=9.0.1",
    "pytest-mock>=3.15.1",
    "ruff>=0.14.5",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
